Organize SDM into models/losses/data/training/evaluation/grounding
Restructure the flat SDM/ package into topical subpackages and update
all relative imports, shell launchers, the smoke test, and the README
layout/usage examples accordingly. Behavior is unchanged; modules now
import as SDM.models.*, SDM.losses.*, SDM.data.*, SDM.training.*,
SDM.evaluation.*, and SDM.grounding.*.
