+++ SDM/models/model.py
@@ -16,7 +16,7 @@ from typing import List
 
 from .encoder import HierarchicalEncoder
 from .predictors import PredictorP3, PredictorP2, CrossStagePredictor, DecoderHead
-from .chest2vec.chest2vec import (
+from ..chest2vec.chest2vec import (
     Chest2Vec, encode_with_eos_ids, get_last_hidden_state,
     last_token_pool, build_qwen_query,
 )
